Modularise create DTO
