+++ packages/amplication-data-service-generator/src/resource/dto/class-transformer.util.ts
@@ -1,0 +1,4 @@
+import { builders } from "ast-types";
+
+export const CLASS_TRANSFORMER_MODULE = "class-transformer";
+export const TYPE_ID = builders.identifier("Type");
