+++ packages/amplication-data-service-generator/src/resource/dto/create-where-unique-input.ts
@@ -1,0 +1,28 @@
+import { builders, namedTypes } from "ast-types";
+import { Entity, EntityField, EnumDataType } from "../../types";
+import { NamedClassDeclaration } from "../../util/ast";
+import { createFieldClassProperty } from "./create-field-class-property";
+
+export function createWhereUniqueInput(
+  entity: Entity,
+  entityIdToName: Record<string, string>
+): NamedClassDeclaration {
+  const uniqueFields = entity.fields.filter(isUniqueField);
+  const properties = uniqueFields.map((field) =>
+    createFieldClassProperty(field, false, true, entityIdToName)
+  );
+  return builders.classDeclaration(
+    createWhereUniqueInputID(entity.name),
+    builders.classBody(properties)
+  ) as NamedClassDeclaration;
+}
+
+export function createWhereUniqueInputID(
+  entityName: string
+): namedTypes.Identifier {
+  return builders.identifier(`${entityName}WhereUniqueInput`);
+}
+
+export function isUniqueField(field: EntityField): boolean {
+  return field.dataType === EnumDataType.Id;
+}
